Restructure and add setup.py
Restructure files and add setup.py to enable upload to pypi.
