+++ lenddo_api_client/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+from lenddo_api_client.base import LenddoAPIClient
+
+
+__all__ = ['LenddoAPIClient']
